requirements.txt
flickrapi
azure-ai-vision